# Add CPU time measurements to Benchmark.ips output