Wall-clock time is a fantastic approximation of scalability if the CPU is never yielded inside the block. Unfortunately, in measuring execution time for database queries, HTTP requests, or things like that, you need a _lot_ of iterations over the block to understand how well it scales in terms of the application. `Benchmark.ips` gives us a really nice way to measure time spent per iteration without having to dial in the number of iterations ourselves, but we can't get CPU usage information from it.

This PR adds that to the `Benchmark.ips` report:

```
➜  redis git:(master) ✗ crystal run --release bench/pipeline_vs_raw.cr
     raw 207.99  (  4.81ms real, 528.26µs user,   1.14ms sys) (± 2.80%)  2.34kB/op  20.77× slower
pipeline   4.32k (231.53µs real,  40.31µs user,  30.75µs sys) (± 0.98%)  21.6kB/op        fastest
```

This will probably need a little more work (for example, I used the simple form of `humanize` instead of the block form as was used elsewhere, which may not be ideal), but I'm curious what people's thoughts are before putting more time into this. 🙂 

If folks are good with this, since we'd be outputting so much information here in a quasi-tabular format, would it make sense to make it an actual table?

```
➜  redis git:(master) ✗ crystal run --release bench/pipeline_vs_raw.cr
Benchmark |     ips |    real |     user |   system | variance |       heap |        result
----------|---------|---------|----------|----------|----------|------------|--------------
     raw  | 207.99  |  4.81ms | 528.26µs |   1.14ms |  ± 2.80% |  2.34kB/op | 20.77× slower
pipeline  |   4.32k |231.53µs |  40.31µs |  30.75µs |  ± 0.98% |  21.6kB/op |       fastest
```