@straight-shoota 2020-10-30
I like the idea 👍 
However, I'm a bit sceptical of the amount of data this prins to the terminal. Having short and concise output is always preferred. So the question is, whether the additional data is relevant enough in a large number of use cases to warrant more convoluted output.
Maybe this could be optional? Also the table formatting.

It's also worth considering that inserting additional time metrics at the proposed location disturbes the visual correlation between variance and the data it's based on (ips / real time).