@asterite 2020-10-30
> Maybe this could be optional?

I also like this, but I think it should be enabled by passing an argument to the `ips` method.