@bcardiff 2020-10-30
FYI, the `Crystal::TablePrint` helper might be helpful here.