@jgaskins 2020-11-13
[Apologies for the delay in response](https://twitter.com/Marissa__Miller/status/703085321643278336). Been a wild few weeks.

@straight-shoota 

> So the question is, whether the additional data is relevant enough in a large number of use cases to warrant more convoluted output.

Whether it's useful in _enough_ cases, I'm not sure. Anecdotally, I work primarily on distributed systems so I/O is a huge component of what I do, and for anything that deals with I/O (even to `localhost`) wall-clock time remains a good latency metric but its usefulness for understanding throughput is reduced to very little as soon as concurrency comes into play.

With that said, though, I think it makes sense to keep `ips` as iterations per wall-clock second. It's still a good baseline for how long the person running the benchmark should expect to wait. Iterations per CPU second in an I/O-heavy scenario could take hours. 🙂

Can you expand on what you think is convoluted, though? I think we may have different interpretations of that word in mind.

---

@asterite 

> I think it should be enabled by passing an argument to the `ips` method.

I think it makes a lot of sense to put the distinction between `user` and `system` CPU time behind a flag since that's a very specific type of insight. But the CPU-time metric can be just as useful as the heap-allocations metric included in the output by default.

Part of this comes from a concern about the ways our tools influence how we measure. If we ignore CPU time by default, I wonder if anyone will _ever_ look at it.

---

@bcardiff 

> FYI, the `Crystal::TablePrint` helper might be helpful here.

Whoa, there's already a thing for this?? If we end up going this route, would it make sense to document that helper/move it out of the `Crystal` namespace or keep it there since this is an internal thing?