@straight-shoota 2020-11-13
> Can you expand on what you think is convoluted, though? I think we may have different interpretations of that word in mind.

I mean that the number of data points per row in `Benchmark.ips` output nearly doubles from four to seven and most uses will probably be irritated about what to do with such an overload of information.

Additionally, the line length grows enormously. In your examples it's 98 characters. Table format it's a bit less because you can leave the units, but not by much. But with so much data, chances are that it might cause the line to wrap if the terminal is not wide enough. Broken lines are going to look terrible. That may be acceptable if you really need the additional data. But I assume most use cases don't.

Bottom line, adding more measurements makes the output harder to read because it presents data most users likely didn't even ask for.